requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
Pillow>=10.2.0
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
matplotlib>=3.8.0
scikit-learn>=1.4.0
pandas>=2.2.0